Do not expose unnecessary entities
