--- core/server/types/next_token_info.go
@@ -1,7 +1,0 @@
-package types
-
-type PageTokenInfo struct {
-	NamespaceIndex int    `json:"namespace_index"`
-	Namespace      string `json:"namespace"`
-	K8sPageToken   string `json:"k8s_page_token"`
-}
